moved magpie_simple to magpie_slashbox, and replaced it with a simpler demo.
